Switched to animate element and object oriented style
This also removes all external javascript/css, everything is now inline.
The animate element is from SMIL/SVG and is used to speficy animations in
markup. This combined with some JavaScript makes it trivial to move the orb
around. The pulsing is done i a similar fashin, with repeatCount used to make
the idle animation trival. However in order to end the animation as it loops
care has to be taken. Lastly a WebSocket is wired up to sync with the server.
All time syncing is done using Promises.